Don't use fjs algorithm if the pattern length is 1

+++ include/fjs.h
@@ -5,6 +5,6 @@
 
 void prepare_fjs(const char *pattern, int pattern_len, enum file_type t);
 void free_fjs();
-bool fjs(const char *buf, int search_len, const char *pattern, int pattern_len, enum file_type t, match *mt);
+bool fjs(const char *buf, size_t search_len, const char *pattern, int pattern_len, enum file_type t, match *mt);
 
 #endif // _HW_FJS_H_
